処理移動 / 暗号化済認証情報ファイルの読み込み機能を AnyDBConnectionStrings のコンストラクタへ移動
同処理は元々、ConnectionDialogViewModel で使用していたもの。
MainWindowViewModel でもこの処理を使用したいので移動する。
